+++ ingest/PUMS/PUMS_data.py
@@ -3,7 +3,7 @@ import pandas as pd
 import numpy as np
 from sqlalchemy import func
 
-from utils.assign_PUMA import assign_PUMA, puma_to_borough
+from utils.PUMA_helpers import assign_PUMA, puma_to_borough
 
 """To do: make this central module from which all other code is called. Write 
 class method for aggregate step to access.  Class method will return cached data or
